fix warnings

+++ modules/imgproc/src/hough.cpp
@@ -2302,7 +2302,7 @@ static void HoughCircles( InputArray _image, OutputArray _circles,
                 for( i = 0; i < ncircles; i++ )
                     cw[i] = GetCircle4f(circles[i]);
                 if (ncircles > 0)
-                    Mat(1, ncircles, cv::traits::Type<Vec4f>::value, &cw[0]).copyTo(_circles);
+                    Mat(1, (int)ncircles, cv::traits::Type<Vec4f>::value, &cw[0]).copyTo(_circles);
             }
             else if( type == CV_32FC3 )
             {
@@ -2310,7 +2310,7 @@ static void HoughCircles( InputArray _image, OutputArray _circles,
                 for( i = 0; i < ncircles; i++ )
                     cwow[i] = GetCircle(circles[i]);
                 if (ncircles > 0)
-                    Mat(1, ncircles, cv::traits::Type<Vec3f>::value, &cwow[0]).copyTo(_circles);
+                    Mat(1, (int)ncircles, cv::traits::Type<Vec3f>::value, &cwow[0]).copyTo(_circles);
             }
             else
                 CV_Error(Error::StsError, "Internal error");
